refactor: reorganize project
- all extension code is together in one place in `src`
- all images are in `media`
- all scripts are in `scripts` with their own eslint config to avoid
  having `eslint-disable` comments all over the place
- removed unused file duplicate.tiff
- removed unused PR ignores for tiff and svg files from codeql workflow
