+++ .vscodeignore
@@ -8,12 +8,12 @@
 
 !CHANGELOG.md
 !README.md
-!screenshot.png
 !LICENSE
 
 # Assets
 
 !media/stylelint.png
+!media/screenshot.png
 
 # Manifest
