+++ scripts/.eslintrc.js
@@ -1,0 +1,12 @@
+'use strict';
+
+/** @type {import('eslint').Linter.Config} */
+const config = {
+	rules: {
+		'no-console': 'off',
+		'no-process-exit': 'off',
+		'node/no-unpublished-require': 'off',
+	},
+};
+
+module.exports = config;
