+++ scripts/ci-download-vscode.js
@@ -1,7 +1,6 @@
 'use strict';
 
-// eslint-disable-next-line node/no-unpublished-require
 const { downloadAndUnzipVSCode } = require('@vscode/test-electron');
-const { version, platform } = require('./jest-runner-vscode.config');
+const { version, platform } = require('../jest-runner-vscode.config');
 
 downloadAndUnzipVSCode(version, platform);
