+++ scripts/download-types.js
@@ -3,20 +3,15 @@
 // Downloads the types file from the stylelint repo to the stylelint-types directory.
 // Overwrites the existing file if it exists.
 
-/* eslint-disable eslint-comments/disable-enable-pair */
-/* eslint-disable no-console */
-/* eslint-disable no-process-exit */
-/* eslint-disable node/no-unpublished-require */
-
 const fs = require('fs-extra');
 const https = require('https');
 const path = require('path');
 
-const typesDir = path.join(__dirname, 'types/stylelint');
+const typesDir = path.join(__dirname, '../types/stylelint');
 const typesPath = path.join(typesDir, 'index.d.ts');
 const typesVersionPath = path.join(typesDir, 'version');
 
-const lockfile = require('./package-lock.json');
+const lockfile = require('../package-lock.json');
 
 const { version } = lockfile?.dependencies?.stylelint;
