+++ src/warnings-to-diagnostics.js
