+++ tsconfig.json
@@ -14,5 +14,5 @@
 		"resolveJsonModule": true,
 		"typeRoots": ["./types", "./node_modules/@types"]
 	},
-	"include": ["server.js", "index.js", "lib"]
+	"include": ["src/**/*"]
 }
